requirements.txt
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.12.3
bech32==1.2.0
Brotli==1.1.0
build==1.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==4.4.2
emoji==2.12.1
executing==2.0.1
ffmpeg==1.4
ffmpegio==0.9.1
ffmpegio-core==0.10.0
idna==3.7
imageio==2.33.1
imageio-ffmpeg==0.4.9
instaloader==4.10.1
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
lnurl==0.4.2
matplotlib-inline==0.1.7
moviepy==2.0.0.dev2
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.3
nostr-dvm==0.8.2
nostr-sdk==0.34.0
numpy==1.26.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.2
pluggy==1.3.0
proglog==0.1.10
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==1.10.13
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
PyUpload==0.1.4
pyzmq==26.0.3
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.13.1
semantic-version==2.9.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.1
tzdata==2023.4
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
yt-dlp==2024.5.27
zipp==3.17.0
cashu~=0.16.0
httpx~=0.25.2